pyproject.toml
[tool.poetry]
name = "eagle-200-collector"
version = "0.1.0"
description = ""
authors = ["Lukas Rusak <lorusak@gmail.com>"]
readme = "README.md"
packages = [{include = "eagle-200-collector", from = "src"}]
include = ["static", "templates"]

[tool.poetry.dependencies]
python = "^3.12"
py-eagle-200 = {git = "https://github.com/lrusak/py-eagle-200.git", rev = "0.1.0"}
aiohttp = "^3.9.5"
python-socketio = "^5.11.2"
aiohttp-jinja2 = "^1.6"
jinja2 = "^3.1.4"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
eagle-200-collector = "eagle-200-collector.main:main"
